+++ packages/atomic/src/components/search/facets/atomic-color-facet/atomic-color-facet.tsx
@@ -71,13 +71,14 @@ import {Bindings} from '../../atomic-search-interface/atomic-search-interface';
  * @part value-* - The dynamic part name used to customize a facet value. The `*` is a syntactical placeholder for a specific facet value. For example, if the component's `field` property is set to 'filetype' and your source has a `YouTubeVideo` file type, the part would be targeted like this: `atomic-color-facet::part(value-YouTubeVideo)...`.
  *
  * @part value-box - The facet value when display is 'box'.
+ * @part value-box-selected - The selected facet value when display is 'box'.
+ * @part value-checkbox - The facet value checkbox, available when display is 'checkbox'.
+ * @part value-checkbox-checked - The checked facet value checkbox, available when display is 'checkbox'.
  * @part value-checkbox-label - The facet value checkbox clickable label, available when display is 'checkbox'.
  *
  * @part show-more - The show more results button.
  * @part show-less - The show less results button.
  * @part show-more-less-icon - The icons of the show more & show less buttons.
- *
- * @part ripple - The ripple effect of the component's interactive elements.
  */
 @Component({
   tag: 'atomic-color-facet',
@@ -90,6 +91,7 @@ export class AtomicColorFacet
   @InitializeBindings() public bindings!: Bindings;
   public facet!: Facet;
   private dependenciesManager?: FacetConditionsManager;
+  private resultIndexToFocusOnShowMore = 0;
   public searchStatus!: SearchStatus;
   @Element() private host!: HTMLElement;
 
@@ -177,7 +179,10 @@ export class AtomicColorFacet
   @MapProp() @Prop() public dependsOn: Record<string, string> = {};
 
   @FocusTarget()
-  private showMoreLessFocus!: FocusTargetController;
+  private showLessFocus!: FocusTargetController;
+
+  @FocusTarget()
+  private showMoreFocus!: FocusTargetController;
 
   @FocusTarget()
   private headerFocus!: FocusTargetController;
@@ -294,7 +299,8 @@ export class AtomicColorFacet
   private renderValue(
     facetValue: FacetValue,
     onClick: () => void,
-    isShowMoreLessFocusTarget: boolean
+    isShowLessFocusTarget: boolean,
+    isShowMoreFocusTarget: boolean
   ) {
     const displayValue = getFieldValueCaption(
       this.facetId!,
@@ -314,10 +320,10 @@ export class AtomicColorFacet
             i18n={this.bindings.i18n}
             onClick={onClick}
             searchQuery={this.facetState.facetSearch.query}
-            buttonRef={(element) =>
-              isShowMoreLessFocusTarget &&
-              this.showMoreLessFocus.setTarget(element)
-            }
+            buttonRef={(element) => {
+              isShowLessFocusTarget && this.showLessFocus.setTarget(element);
+              isShowMoreFocusTarget && this.showMoreFocus.setTarget(element);
+            }}
           >
             <FacetValueLabelHighlight
               displayValue={displayValue}
@@ -335,10 +341,10 @@ export class AtomicColorFacet
             i18n={this.bindings.i18n}
             onClick={onClick}
             searchQuery={this.facetState.facetSearch.query}
-            buttonRef={(element) =>
-              isShowMoreLessFocusTarget &&
-              this.showMoreLessFocus.setTarget(element)
-            }
+            buttonRef={(element) => {
+              isShowLessFocusTarget && this.showLessFocus.setTarget(element);
+              isShowMoreFocusTarget && this.showMoreFocus.setTarget(element);
+            }}
           >
             <div
               part={`value-${partValueWithDisplayValue} value-${partValueWithAPIValue} default-color-value`}
@@ -374,14 +380,22 @@ export class AtomicColorFacet
   private renderValues() {
     return this.renderValuesContainer(
       this.facetState.values.map((value, i) =>
-        this.renderValue(value, () => this.facet.toggleSelect(value), i === 0)
+        this.renderValue(
+          value,
+          () => this.facet.toggleSelect(value),
+          i === 0,
+          i ===
+            (this.sortCriteria === 'automatic'
+              ? 0
+              : this.resultIndexToFocusOnShowMore)
+        )
       )
     );
   }
 
   private renderSearchResults() {
     return this.renderValuesContainer(
-      this.facetState.facetSearch.values.map((value, i) =>
+      this.facetState.facetSearch.values.map((value) =>
         this.renderValue(
           {
             state: 'idle',
@@ -389,7 +403,8 @@ export class AtomicColorFacet
             value: value.rawValue,
           },
           () => this.facet.facetSearch.select(value),
-          i === 0
+          false,
+          false
         )
       ),
       this.facetState.facetSearch.query
@@ -413,11 +428,12 @@ export class AtomicColorFacet
         label={this.label}
         i18n={this.bindings.i18n}
         onShowMore={() => {
-          this.showMoreLessFocus.focusAfterSearch();
+          this.resultIndexToFocusOnShowMore = this.facet.state.values.length;
+          this.showMoreFocus.focusAfterSearch();
           this.facet.showMoreValues();
         }}
         onShowLess={() => {
-          this.showMoreLessFocus.focusAfterSearch();
+          this.showLessFocus.focusAfterSearch();
           this.facet.showLessValues();
         }}
         canShowLessValues={this.facetState.canShowLessValues}