+++ packages/atomic/src/components/search/facets/atomic-numeric-facet/atomic-numeric-facet.tsx
@@ -73,17 +73,17 @@ interface NumericRangeWithLabel extends NumericRangeRequest {
  * @part value-count - The facet value count, common for all displays.
  *
  * @part value-checkbox - The facet value checkbox, available when display is 'checkbox'.
+ * @part value-checkbox-checked - The checked facet value checkbox, available when display is 'checkbox'.
  * @part value-checkbox-label - The facet value checkbox clickable label, available when display is 'checkbox'.
  * @part value-link - The facet value when display is 'link'.
- *
+ * @part value-link-selected - The selected facet value when display is 'link'.
+
  * @part input-form - The form that comprises the labels, inputs, and 'apply' button for the custom numeric range.
  * @part label-start - The label for the starting value of the custom numeric range.
  * @part label-end - The label for the ending value of the custom numeric range.
  * @part input-start - The input for the starting value of the custom numeric range.
  * @part input-end - The input for the ending value of the custom numeric range.
  * @part input-apply-button - The apply button for the custom range.
- *
- * @part ripple - The ripple effect of the component's interactive elements.
  */
 @Component({
   tag: 'atomic-numeric-facet',