+++ packages/atomic/src/components/search/facets/atomic-rating-facet/atomic-rating-facet.tsx
@@ -54,11 +54,10 @@ import {Bindings} from '../../atomic-search-interface/atomic-search-interface';
  * @part value-rating - The facet value rating, common for all displays.
  *
  * @part value-checkbox - The facet value checkbox, available when display is 'checkbox'.
+ * @part value-checkbox-checked - The checked facet value checkbox, available when display is 'checkbox'.
  * @part value-checkbox-label - The facet value checkbox clickable label, available when display is 'checkbox'.
  * @part value-link - The facet value when display is 'link'.
- *
- * @part ripple - The ripple effect of the component's interactive elements.
- *
+ * @part value-link-selected - The selected facet value when display is 'link'.
  */
 @Component({
   tag: 'atomic-rating-facet',