+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,94 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.25](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.24...@coveo/atomic-react@1.15.25) (2022-07-27)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.24](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.23...@coveo/atomic-react@1.15.24) (2022-07-27)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.23](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.22...@coveo/atomic-react@1.15.23) (2022-07-26)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.22](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.21...@coveo/atomic-react@1.15.22) (2022-07-26)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.21](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.20...@coveo/atomic-react@1.15.21) (2022-07-25)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.20](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.19...@coveo/atomic-react@1.15.20) (2022-07-25)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.19](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.18...@coveo/atomic-react@1.15.19) (2022-07-25)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.18](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.17...@coveo/atomic-react@1.15.18) (2022-07-25)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.17](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.16...@coveo/atomic-react@1.15.17) (2022-07-22)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.16](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.15...@coveo/atomic-react@1.15.16) (2022-07-21)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.15](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.14...@coveo/atomic-react@1.15.15) (2022-07-21)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 ## [1.15.14](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.13...@coveo/atomic-react@1.15.14) (2022-07-15)
 
 **Note:** Version bump only for package @coveo/atomic-react